some file and code structure updates
