+++ src/api/services/authApi.ts
@@ -1,0 +1,11 @@
+import { supabase } from '../supabase';
+
+export const authLogin = async () =>
+  supabase.auth.signInWithOAuth({ provider: 'google' });
+
+export const authSignOut = async () => supabase.auth.signOut();
+
+export const authGetSession = async () => {
+  const { data, error } = await supabase.auth.getSession();
+  return { session: data.session, error };
+};
