+++ src/config/config.ts
@@ -1,0 +1,1 @@
+export const USERS_TABLE = 'app-users';
